requirements.txt
Pillow
opencv-python
pdf2image
scipy
matplotlib